requirements.txt
asgiref==3.5.0
django-cors-headers==3.12.0
django-ratings==0.3.7
djangorestframework==3.13.1
djongo==1.3.6
Pillow==9.1.0
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
